Refactor reducer imports
